Move code related to handling IO address DD00 to where it belongs in CIA chip code.
